rename pattern dfa
