+++ src/parser/Makefile
@@ -2,16 +2,16 @@ CC = cc
 CFLAGS = -Wall -g -D__DEBUG__ -I ../basekit
 VPATH :=../basekit:test
 
-all: sol_dfa.o sol_pattern_dfa.o
+all: sol_dfa.o sol_pattern.o
 
 sol_dfa.o: sol_dfa.c sol_common.h sol_hash.o
-sol_pattern_dfa.o: sol_pattern_dfa.c sol_dfa.o
+sol_pattern.o: sol_pattern.c sol_dfa.o
 
 test_dfa: test_dfa.c sol_dfa.o sol_common.h sol_hash.o sol_utils.o  Hash_fnv.c Hash_murmur.c
 	if [ ! -d output ]; then mkdir output; fi
 	$(CC) $(CFLAGS) -o output/$@ $^
 
-test_pattern_dfa: test_pattern_dfa.c sol_pattern_dfa.o sol_dfa.o sol_hash.o sol_utils.o sol_stack.o sol_list.o Hash_fnv.c Hash_murmur.c
+test_pattern: test_pattern.c sol_pattern.o sol_dfa.o sol_hash.o sol_utils.o sol_stack.o sol_list.o Hash_fnv.c Hash_murmur.c
 	if [ ! -d output ]; then mkdir output; fi
 	$(CC) $(CFLAGS) -o output/$@ $^
