+++ src/parser/sol_pattern.h
